+++ CHANGELOG.md
@@ -1,4 +1,24 @@
+## 0.0.5
+
+*   [change] **break-change**: `[]` means anything in it is optional.
+
+    ```
+    Usage: prog [-dir]
+    ```
+
+    equals to
+
+    ```
+    Usage: prog [-d] [-i] [-r]
+    ```
+
+    If you want all elements to appear, use
+
+    ```
+    Usage: prog [(-dir)]
+    ```
+
 ## 0.0.4
 
-[fix] `Usage: prog [-v] -- <args>...` failed to match `porg -v -- -v --verb`
-[fix] [#1](https://github.com/TylerTemp/docpie/issues/1) command order bug.
+*   [fix] `Usage: prog [-v] -- <args>...` failed to match `porg -v -- -v --verb`
+*   [fix] [#1](https://github.com/TylerTemp/docpie/issues/1) command order bug.